rename from account to session in contexts
