+++ app/ts/components/TransferButton.tsx
@@ -3,10 +3,10 @@ import { JSX } from 'preact/jsx-runtime'
 import { removeNonStringsAndTrim } from '../library/utilities.js'
 import { useTransfer } from '../context/Transfer.js'
 import * as Icon from './Icon/index.js'
-import { useWallet } from '../context/Wallet.js'
+import { useEthereumProvider } from '../context/Ethereum.js'
 import { useAsyncState } from '../library/preact-utilities.js'
 import { EthereumAddress } from '../schema.js'
-import { useAccount } from '../context/Account.js'
+import { useWallet } from '../context/Wallet.js'
 import { makeError } from 'ethers'
 
 export const TransferButton = () => {
@@ -27,8 +27,8 @@ export const TransferButton = () => {
 
 const ConnectOrTransferButton = () => {
 	const { value: query, waitFor } = useAsyncState<EthereumAddress>()
-	const { browserProvider } = useWallet()
-	const { account } = useAccount()
+	const { browserProvider } = useEthereumProvider()
+	const { account } = useWallet()
 
 	const connect = () => {
 		waitFor(async () => {
